Added the last step
I was running into an issue where the code kept predicting my large crack to be 100% small. I have added modifications to it.

They are...

Added custom learning rate to Adam, i slowed it down to make it learn correctly

reduced the drop rate, it think it was dropping to much

i added an extra layer with 128 neurons to add complexity to the network

added more epoch iters, 10->20
this seems to work